feat: Add multer for image upload in teamController
The code changes include adding the multer package as a dependency in the `package.json` file. This package is used for handling file uploads, specifically for uploading team photos. The `teamModel.js` file has been modified to include a new field `photo` for storing the URL of the team's photo. The `teams.js` file now includes middleware for handling image uploads using multer. This feature allows users to upload and associate photos with teams in the application.

Recent repository commits:
- chore: Fix typo in requireAuth import
- feat: Add routes for creating academies and users
- feat: Add payment model
- feat: Add user, academy, player, and event models
- chore: Update dotenv dependency to version 16.4.5
- chore: Add initial server setup and dependencies